+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Release 0.12.0 (unreleased)
 * Skip patches outside manifest dir (#942)
 * Make patch path in metadata platform independent (#937)
 * Fix extra newlines in patch for new files (#945)
+* Introduce new ``update-patch`` command (#614)
 
 Release 0.11.0 (released 2026-01-03)
 ====================================